Adds in function for fully loading the cache from a simulation as well as getting the save states to work.
Refactors simulation loading to use a cache, improving efficiency and
reducing API calls during initial setup. This change also enhances error
handling by adding a try-except block to prevent queue-related exceptions
from crashing the client.

Removes the `tick` and `tick_duration` methods. Instead the structure of the
simulation is now loaded in its entireity when the simulation is created to
ensure the state is accurate.

The system now pulls state directly from the endpoint, rather than processing
through the invoke function. This ensures the state is accurate and not
processed or deserialized.